Tidied folder
Removed patches that I did not need any more, mainly the fibonnaci, deck,
smartborders and scratchpad patches. Fib and deck are replaced by vanitygaps as
it contains those layouts already, and smartborders was too long for me to
bother troubleshooting.
The scratchpad patch was replaced by the scratchpadS patch instead. It allows
for multiple scratchpads.